+++ exe/arduino_ci.rb
@@ -24,7 +24,7 @@ class Parser
       ci_config: {
         "unittest" => unit_config
       },
-      min_free_space: 0,
+      min_free_space: nil,
     }
 
     opt_parser = OptionParser.new do |opts|
@@ -50,7 +50,7 @@ class Parser
         unit_config["testfiles"]["reject"] << p
       end
 
-      opts.on("--min-free-space=VALUE", "Minimum free SRAM memory for stack/heap") do |p|
+      opts.on("--min-free-space=VALUE", "Minimum free SRAM memory for stack/heap, in bytes") do |p|
         output_options[:min_free_space] = p.to_i
       end
 
@@ -502,22 +502,19 @@ def perform_example_compilation_tests(cpp_library, config)
       board = ovr_config.platform_info[p][:board]
       attempt("Compiling #{example_name} for #{board}") do
         ret = @backend.compile_sketch(example_path, board)
-        puts
-        if ret
-          output = @backend.last_msg
-          puts output
-          i = output.index("leaving")
-          free_space = output[i + 8..-1].to_i
-          min_free_space = @cli_options[:min_free_space]
-          if free_space < min_free_space
-            puts "Free space of #{free_space} is less than minimum of #{min_free_space}"
-            ret = false
-          end
-        else
+        unless ret
           puts "Last command: #{@backend.last_msg}"
           puts @backend.last_err
         end
         ret
+      end
+
+      next if @cli_options[:min_free_space].nil?
+
+      usage = @backend.last_bytes_usage
+      min_free_space = @cli_options[:min_free_space]
+      attempt("Checking that free space of #{usage[:free]} is less than desired minimum #{min_free_space}") do
+        min_free_space <= usage[:free]
       end
     end
   end